LayoutTests/platform/mac-monterey/imported/w3c/web-platform-tests/fetch/metadata/generated/element-meta-refresh.https.optional.sub-expected.txt


Harness Error (TIMEOUT), message = null

FAIL sec-fetch-site - Same origin assert_array_equals: expected property 0 to be "same-origin" but got "none" (expected array ["same-origin"] got ["none"])
FAIL sec-fetch-site - Cross-site assert_array_equals: expected property 0 to be "cross-site" but got "none" (expected array ["cross-site"] got ["none"])
TIMEOUT sec-fetch-site - Same site Test timed out
NOTRUN sec-fetch-site - Same-Origin -> Cross-Site -> Same-Origin redirect
NOTRUN sec-fetch-site - Same-Origin -> Same-Site -> Same-Origin redirect
NOTRUN sec-fetch-site - Cross-Site -> Same Origin
NOTRUN sec-fetch-site - Cross-Site -> Same-Site
NOTRUN sec-fetch-site - Cross-Site -> Cross-Site
NOTRUN sec-fetch-site - Same-Origin -> Same Origin
NOTRUN sec-fetch-site - Same-Origin -> Same-Site
NOTRUN sec-fetch-site - Same-Origin -> Cross-Site
NOTRUN sec-fetch-site - Same-Site -> Same Origin
NOTRUN sec-fetch-site - Same-Site -> Same-Site
NOTRUN sec-fetch-site - Same-Site -> Cross-Site
NOTRUN sec-fetch-site - HTTPS downgrade-upgrade
NOTRUN sec-fetch-mode
NOTRUN sec-fetch-dest
NOTRUN sec-fetch-user
